refactor (#411)
Structural changes to allow testing the parsing of specs themselves.
